build: always run on-device
On macOS, `adb devices` ALWAYS has an 'offline' emulator device:

```
List of devices attached
emulator-5562  offline
```

As a result, all the `adb` commands in build.zig only work if you are
using the emulator. The second that you connect an actual Android device,
you have two devices in the list - and it all the adb commands error out
with a 'ambiguous device' error.

Adding the `-d` flag everywhere tells adb to prefer physical devices,
which results in build.zig only working with physical devices in
practice.

Long-term, I imagine you would want to make `-d` a CLI flag that can be
passed to `zig build` - or perhaps even let the user select a device
somehow. But this was good enough for my purposes.

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>